// js/classes/types/GlobalConfig.ts
export type GlobalConfig = {
    noiseScale: number;
    noiseOffset: number;
    heightScale: number;
    waterLevel: number;
    castleFlattenDistance: number;
    treeNoiseScale: number;
    treeNoiseOffset: number;
    treeThreshold: number;
    treeSparseThreshold: number;
    rockThreshold: number;
};